# perf: basic criterion benchmarks for store and load of text