--- a/Cargo.toml
+++ b/Cargo.toml
@@ -49,3 +49,8 @@ libipld = "0.8"
 
 [dev-dependencies]
 tempfile = "3.1"
+criterion = "0.3"
+
+[[bench]]
+name = "ipfs"
+harness = false