--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod data;
+pub mod message;
+pub mod page;
+pub mod ui;