@cryptoquick 2021-01-09
Great work also! I'm fine with moving the binary file, and I'll try to make sure everything works well when I do a release. I might have to add a `[[bin]]` section to the Cargo.toml.

Another issue with forking your changes in, as much as I love the model, is that github doesn't run actions over them. Can you push both of your changes to the repo itself? Also, maybe join our Discord. I should put a link up in the README to our Matrix server, also.